Move mypy wrapper to tools
