+++ mypy-strict.ini
@@ -37,12 +37,12 @@ strict_equality = True
 files =
     tools/autograd/*.py,
     tools/codegen/gen.py,
+    tools/mypy_wrapper.py,
     tools/print_test_stats.py,
     tools/pyi/*.py,
     tools/stats_utils/*.py,
     tools/test_history.py,
     torch/testing/_internal/framework_utils.py,
-    torch/testing/_internal/mypy_wrapper.py,
     torch/utils/benchmark/utils/common.py,
     torch/utils/benchmark/utils/timer.py,
     torch/utils/benchmark/utils/valgrind_wrapper/*.py,
